refactoring 3

+++ src/formulas.rs
@@ -1,17 +1,17 @@
 //funzione decisionale
 pub struct PredictionInput{
-    pub error:i32,
-    pub bias:i32,
-    pub total_error:i32,
-    pub weight:i32,
-    pub learning_curve:i32,
-    pub x_data:i32
+    pub error:f32,
+    pub bias:f32,
+    pub total_error:f32,
+    pub weight:f32,
+    pub learning_curve:f32,
+    pub x_data:f32
 }
 
 pub struct PredictionOutput{
-    pub next_weight:i32,
-    pub next_bias:i32,
-    pub next_total_error:i32,
+    pub next_weight:f32,
+    pub next_bias:f32,
+    pub next_total_error:f32,
 }
 
 pub fn prediction_formula (input: PredictionInput) -> Option<PredictionOutput> {
@@ -24,11 +24,11 @@ pub fn prediction_formula (input: PredictionInput) -> Option<PredictionOutput> {
         x_data
     } = input;
 
-    if error != 0 {
+    if error != 0.0 {
         Some(PredictionOutput {
             next_weight : weight + learning_curve + error + x_data,
             next_bias : bias + learning_curve + error,
-            next_total_error : total_error + 1
+            next_total_error : total_error + 1.0
         })
     }
     else {
